c4_rust_Al_Jazirah_Al_Hamra::vm
Struct VM

pub struct VM {
    pub text: Vec<Instruction>,
    pub stack: Vec<i32>,
    pub pc: usize,
    pub bp: usize,
    pub sp: usize,
    pub ax: i32,
    pub variables: HashMap<String, i32>,
    pub functions: HashMap<String, Function>,
    pub call_stack: Vec<usize>,
    pub variable_stack: Vec<HashMap<String, i32>>,
}

Virtual machine that executes compiled instructions

Fields
text: Vec<Instruction>stack: Vec<i32>pc: usizebp: usizesp: usizeax: i32variables: HashMap<String, i32>functions: HashMap<String, Function>call_stack: Vec<usize>variable_stack: Vec<HashMap<String, i32>>
Implementations

impl VM

pub fn new( program: Vec<Instruction>, functions: HashMap<String, Function>, ) -> Self

Creates a new VM with given program instructions and function definitions

pub fn run(&mut self) -> i32

Runs the program from start to finish and returns the final result

fn execute_instruction(&mut self)

fn fetch(&mut self) -> Instruction

fn exec_imm(&mut self, val: i32)

fn exec_push(&mut self)

fn exec_add(&mut self)

fn exec_sub(&mut self)

fn exec_mul(&mut self)

fn exec_div(&mut self)

fn exec_mod(&mut self)

fn exec_gt(&mut self)

fn exec_lt(&mut self)

fn exec_eq(&mut self)

fn exec_ne(&mut self)

fn exec_jz(&mut self, addr: usize)

fn exec_jmp(&mut self, addr: usize)

fn exec_ent(&mut self, size: usize)

fn exec_adj(&mut self, size: usize)

fn exec_lev(&mut self)

fn exec_lea(&mut self, offset: usize)

fn exec_call(&mut self, name: &str)

fn exec_return(&mut self)

fn exec_store(&mut self, name: &str)

fn exec_load(&mut self, name: String)

pub fn exec_printf(&mut self, fmt: &String, args: &Vec<String>)

Executes a printf instruction with format string and arguments

fn exec_load_string(&mut self, string: String)

fn exec_deref(&mut self)

fn exec_addr(&mut self, name: &str)